// resources/js/lab/runtime/src/sdk/types.ts
export interface StatResult {
  isFile: boolean;
  isDirectory: boolean;
  size: number;
  mtime: number;
}